+++ P2P/FloatMath.cs
@@ -1,0 +1,19 @@
+namespace P2P;
+
+public static class FloatMath
+{
+	public static float Sin(double value)
+	{
+		return (float)Math.Sin(value);
+	}
+	
+	public static float Cos(double value)
+	{
+		return (float)Math.Cos(value);
+	}
+
+	public static float Sqrt(double value)
+	{
+		return (float)Math.Sqrt(value);
+	}
+}
